:sparkles: adds results sending and receiving logic to agents

+++ src/main/java/hierarchical_delegation/strategies/SortStrategy.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class SortStrategy implements Strategy {
     @Override
-    public ArrayList<Integer> executeOperation(ArrayList<Integer> data) {
+    public ArrayList<Double> executeOperation(ArrayList<Double> data) {
         data.sort(null);
         return data;
     }
